Rename abstract classes to `Base`, method shuffling
