+++ src/main/java/frc/robot/commands/sequences/BlueTopAuton.java
@@ -12,7 +12,7 @@ import frc.robot.subsystems.drivetrain.BaseSwerveSubsystem;
  * The top and bottom auton sequences, where the robot deposits its preloaded game piece and picks up
  * and places another.
  */
-public class BlueTopAuton extends TopAutonSequence {
+public class BlueTopAuton extends BaseTopAutonSequence {
     private static final Pose2d INITIAL_POSE = new Pose2d(
         Units.inchesToMeters(PlacePosition.BLUE_INIT_X_IN),
         Units.inchesToMeters(195.55),
